backend/requirements.txt
requests
beautifulsoup4
fastapi
uvicorn
python-jose[cryptography]
pydantic
python-jose
cryptography
passlib==1.7.4
bcrypt==3.2.0
python-multipart
python-dotenv
SQLAlchemy
psycopg2-binary
alembic
bandit
